Abstract drugs and chemical probes into directories
